+++ .github/workflows/certification.yml
@@ -50,7 +50,7 @@ jobs:
         PR_COMPONENTS=$(yq -I0 --tojson eval - << EOF
         - pubsub.kafka
         - pubsub.rabbitmq
-        - pubsub.mqtt
+        - pubsub.mqtt3
         - state.mongodb
         - state.redis
         - state.cockroachdb
