+++ .github/workflows/conformance.yml
@@ -53,9 +53,9 @@ jobs:
         - bindings.influx
         - bindings.kafka-wurstmeister
         - bindings.kafka-confluent
-        - bindings.mqtt-emqx
-        - bindings.mqtt-mosquitto
-        - bindings.mqtt-vernemq
+        - bindings.mqtt3-emqx
+        - bindings.mqtt3-mosquitto
+        - bindings.mqtt3-vernemq
         - bindings.postgres
         - bindings.redis.v6
         - bindings.redis.v7
@@ -66,9 +66,9 @@ jobs:
         - configuration.redis.v7
         - pubsub.hazelcast
         - pubsub.in-memory
-        - pubsub.mqtt-emqx
-        - pubsub.mqtt-mosquitto
-        - pubsub.mqtt-vernemq
+        - pubsub.mqtt3-emqx
+        - pubsub.mqtt3-mosquitto
+        - pubsub.mqtt3-vernemq
         - pubsub.natsstreaming
         - pubsub.pulsar
         - pubsub.rabbitmq
