+++ bindings/mqtt3/metadata.go
