+++ tests/certification/pubsub/mqtt3/components/consumer2/mqtt.yaml
@@ -1,19 +1,19 @@
 apiVersion: dapr.io/v1alpha1
 kind: Component
 metadata:
-  name: pubsub
+  name: messagebus
 spec:
-  type: pubsub.mqtt
+  type: pubsub.mqtt3
   version: v1
   metadata:
   - name: url
-    value: "tcp://localhost:1883"
+    value: "tcp://localhost:1884"
   - name: consumerID
-    value: "testConsumer"
+    value: "testConsumer2"
   - name: retain
     value: false
   - name: qos
-    value: 1
+    value: 2
   - name: cleanSession
     value: false
   - name: backOffMaxRetries
