+++ tests/certification/pubsub/mqtt3/components/consumer3/mqtt.yaml
@@ -3,7 +3,7 @@ kind: Component
 metadata:
   name: messagebus
 spec:
-  type: pubsub.mqtt
+  type: pubsub.mqtt3
   version: v1
   metadata:
   - name: url
