+++ tests/config/bindings/mqtt/emqx/bindings.yml
@@ -3,7 +3,7 @@ kind: Component
 metadata:
   name: binding-mqtt-emqx
 spec:
-  type: bindings.mqtt
+  type: bindings.mqtt3
   version: v1
   metadata:
   - name: url
