+++ tests/config/bindings/mqtt/mosquitto/bindings.yml
@@ -3,7 +3,7 @@ kind: Component
 metadata:
   name: binding-mqtt-mosquitto
 spec:
-  type: bindings.mqtt
+  type: bindings.mqtt3
   version: v1
   metadata:
   - name: url
