+++ tests/config/pubsub/mqtt3/mosquitto/pubsub.yml
@@ -1,19 +1,19 @@
 apiVersion: dapr.io/v1alpha1
 kind: Component
 metadata:
-  name: messagebus
+  name: pubsub
 spec:
-  type: pubsub.mqtt
+  type: pubsub.mqtt3
   version: v1
   metadata:
   - name: url
-    value: "tcp://localhost:1884"
+    value: "tcp://localhost:1883"
   - name: consumerID
-    value: "testConsumer2"
+    value: "testConsumer"
   - name: retain
     value: false
   - name: qos
-    value: 2
+    value: 1
   - name: cleanSession
     value: false
   - name: backOffMaxRetries
