+++ tests/config/pubsub/mqtt3/vernemq/pubsub.yml
@@ -3,11 +3,11 @@ kind: Component
 metadata:
   name: pubsub
 spec:
-  type: pubsub.mqtt
+  type: pubsub.mqtt3
   version: v1
   metadata:
   - name: url
-    value: "tcp://localhost:1884"
+    value: "tcp://localhost:1885"
   - name: consumerID
     value: "testConsumer"
   - name: retain
