+++ tests/config/pubsub/tests.yml
@@ -63,15 +63,15 @@ components:
     allOperations: true
   - component: pulsar
     operations: ['publish', 'subscribe', 'multiplehandlers']
-  - component: mqtt
+  - component: mqtt3
     profile: mosquitto
     operations: ['publish', 'subscribe', 'multiplehandlers']
   - component: solace.amqp
     operations: ['publish', 'subscribe']
-  - component: mqtt
+  - component: mqtt3
     profile: emqx
     operations: ['publish', 'subscribe', 'multiplehandlers']
-  - component: mqtt
+  - component: mqtt3
     profile: vernemq
     operations: ['publish', 'subscribe', 'multiplehandlers']
   - component: hazelcast
